backend/pyproject.toml
[tool.poetry]
name = "rericom_test_task"
version = "0.1.0"
description = ""
authors = ["Timur <phoenixwq01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.0"
djangorestframework = "^3.13.1"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
